port variable should be unsigned short

+++ src/net/py_load_http_connector.cpp
@@ -8,7 +8,7 @@
 
 
 PyLoadHttpConnector::PyLoadHttpConnector(const std::string& hostname,
-                                         const int port) :
+                                         const unsigned short port) :
     url("http://" + hostname + ":" + util::intToString(port) + "/api/")
 {
 }
